requirements.txt
asgiref==3.9.1
bcrypt==4.3.0
Django==5.2.5
django-filter==25.2
django-tinymce==5.0.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
mysqlclient==2.2.7
PyJWT==2.10.1
python-dotenv==1.1.1
sqlparse==0.5.3
tzdata==2025.2 
